Lift libvcx related code from aries_vcx to libvcx_core
