--- aries_vcx/src/handlers/issuance/mediated_holder.rs
@@ -1,69 +1,0 @@
-use std::collections::HashMap;
-
-use messages::{
-    msg_fields::protocols::{
-        cred_issuance::{v1::CredentialIssuanceV1, CredentialIssuance},
-        notification::Notification,
-    },
-    AriesMessage,
-};
-
-use crate::{
-    handlers::{
-        issuance::holder::Holder,
-        util::{matches_opt_thread_id, matches_thread_id},
-    },
-    protocols::issuance::holder::state_machine::HolderState,
-};
-
-pub fn holder_find_message_to_handle(
-    sm: &Holder,
-    messages: HashMap<String, AriesMessage>,
-) -> Option<(String, AriesMessage)> {
-    trace!("holder_find_message_to_handle >>>");
-    for (uid, message) in messages {
-        match sm.get_state() {
-            HolderState::ProposalSet => {
-                if let AriesMessage::CredentialIssuance(CredentialIssuance::V1(
-                    CredentialIssuanceV1::OfferCredential(offer),
-                )) = &message
-                {
-                    if matches_opt_thread_id!(offer, sm.get_thread_id().unwrap().as_str()) {
-                        return Some((uid, message));
-                    }
-                }
-            }
-            HolderState::RequestSet => match &message {
-                AriesMessage::CredentialIssuance(CredentialIssuance::V1(
-                    CredentialIssuanceV1::IssueCredential(credential),
-                )) => {
-                    if matches_thread_id!(credential, sm.get_thread_id().unwrap().as_str()) {
-                        return Some((uid, message));
-                    }
-                }
-                AriesMessage::CredentialIssuance(CredentialIssuance::V1(
-                    CredentialIssuanceV1::ProblemReport(problem_report),
-                )) => {
-                    if matches_opt_thread_id!(problem_report, sm.get_thread_id().unwrap().as_str())
-                    {
-                        return Some((uid, message));
-                    }
-                }
-                AriesMessage::ReportProblem(problem_report) => {
-                    if matches_opt_thread_id!(problem_report, sm.get_thread_id().unwrap().as_str())
-                    {
-                        return Some((uid, message));
-                    }
-                }
-                AriesMessage::Notification(Notification::ProblemReport(msg)) => {
-                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
-                        return Some((uid, message));
-                    }
-                }
-                _ => {}
-            },
-            _ => {}
-        };
-    }
-    None
-}
