+++ aries_vcx/src/handlers/issuance/mod.rs
@@ -1,4 +1,2 @@
 pub mod holder;
 pub mod issuer;
-pub mod mediated_holder;
-pub mod mediated_issuer;
