+++ aries_vcx/src/handlers/proof_presentation/mod.rs
@@ -1,5 +1,3 @@
-pub mod mediated_prover;
-pub mod mediated_verifier;
 pub mod prover;
 pub mod types;
 pub mod verifier;
