+++ libvcx_core/src/api_vcx/api_handle/disclosed_proof.rs
@@ -1,7 +1,6 @@
+use std::collections::HashMap;
 use aries_vcx::{
-    handlers::proof_presentation::{
-        mediated_prover::prover_find_message_to_handle, prover::Prover,
-    },
+    handlers::proof_presentation::prover::Prover,
     messages::{
         msg_fields::protocols::present_proof::{
             v1::{request::RequestPresentationV1, PresentProofV1},
@@ -10,7 +9,10 @@ use aries_vcx::{
         AriesMessage,
     },
 };
+use aries_vcx::handlers::util::{matches_opt_thread_id, matches_thread_id};
 use serde_json;
+use aries_vcx::messages::msg_fields::protocols::notification::Notification;
+use aries_vcx::protocols::proof_presentation::prover::state_machine::ProverState;
 
 use super::mediated_connection::send_message;
 use crate::{
@@ -89,6 +91,69 @@ pub fn get_state(handle: u32) -> LibvcxResult<u32> {
         .map_err(|e| {
             LibvcxError::from_msg(LibvcxErrorKind::InvalidDisclosedProofHandle, e.to_string())
         })
+}
+
+pub fn prover_find_message_to_handle(
+    sm: &Prover,
+    messages: HashMap<String, AriesMessage>,
+) -> Option<(String, AriesMessage)> {
+    trace!("prover_find_message_to_handle >>> messages: {:?}", messages);
+    for (uid, message) in messages {
+        match sm.get_state() {
+            ProverState::PresentationProposalSent => match &message {
+                AriesMessage::ReportProblem(msg) => {
+                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::Notification(Notification::ProblemReport(msg)) => {
+                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::PresentProof(PresentProof::V1(
+                                               PresentProofV1::RequestPresentation(msg),
+                                           )) => {
+                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                _ => {}
+            },
+            ProverState::PresentationSent => match &message {
+                AriesMessage::Notification(Notification::Ack(msg)) => {
+                    if matches_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::PresentProof(PresentProof::V1(PresentProofV1::Ack(msg))) => {
+                    if matches_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::ReportProblem(msg) => {
+                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::Notification(Notification::ProblemReport(msg)) => {
+                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::PresentProof(PresentProof::V1(PresentProofV1::ProblemReport(
+                                                                msg,
+                                                            ))) => {
+                    if matches_opt_thread_id!(msg, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                _ => {}
+            },
+            _ => {}
+        };
+    }
+    None
 }
 
 pub async fn update_state(
