+++ libvcx_core/src/api_vcx/api_handle/proof.rs
@@ -1,15 +1,18 @@
+use std::collections::HashMap;
 use aries_vcx::{
     common::proofs::proof_request::PresentationRequestData,
-    handlers::proof_presentation::{
-        mediated_verifier::verifier_find_message_to_handle, verifier::Verifier,
-    },
+    handlers::proof_presentation::verifier::Verifier,
     messages::AriesMessage,
     protocols::{
         proof_presentation::verifier::verification_status::PresentationVerificationStatus,
         SendClosure,
     },
 };
+use aries_vcx::handlers::util::{matches_opt_thread_id, matches_thread_id};
 use serde_json;
+use aries_vcx::messages::msg_fields::protocols::present_proof::PresentProof;
+use aries_vcx::messages::msg_fields::protocols::present_proof::v1::PresentProofV1;
+use aries_vcx::protocols::proof_presentation::verifier::state_machine::VerifierState;
 
 use crate::{
     api_vcx::{
@@ -54,6 +57,58 @@ pub async fn create_proof(
 
 pub fn is_valid_handle(handle: u32) -> bool {
     PROOF_MAP.has_handle(handle)
+}
+
+pub fn verifier_find_message_to_handle(
+    sm: &Verifier,
+    messages: HashMap<String, AriesMessage>,
+) -> Option<(String, AriesMessage)> {
+    trace!(
+        "verifier_find_message_to_handle >>> messages: {:?}",
+        messages
+    );
+    for (uid, message) in messages {
+        match sm.get_state() {
+            VerifierState::Initial => match &message {
+                AriesMessage::PresentProof(PresentProof::V1(
+                                               PresentProofV1::ProposePresentation(_),
+                                           )) => {
+                    return Some((uid, message));
+                }
+                AriesMessage::PresentProof(PresentProof::V1(
+                                               PresentProofV1::RequestPresentation(_),
+                                           )) => {
+                    return Some((uid, message));
+                }
+                _ => {}
+            },
+            VerifierState::PresentationRequestSent => match &message {
+                AriesMessage::PresentProof(PresentProof::V1(PresentProofV1::Presentation(
+                                                                presentation,
+                                                            ))) => {
+                    if matches_thread_id!(presentation, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::PresentProof(PresentProof::V1(
+                                               PresentProofV1::ProposePresentation(proposal),
+                                           )) => {
+                    if matches_opt_thread_id!(proposal, sm.get_thread_id().unwrap().as_str()) {
+                        return Some((uid, message));
+                    }
+                }
+                AriesMessage::ReportProblem(problem_report) => {
+                    if matches_opt_thread_id!(problem_report, sm.get_thread_id().unwrap().as_str())
+                    {
+                        return Some((uid, message));
+                    }
+                }
+                _ => {}
+            },
+            _ => {}
+        };
+    }
+    None
 }
 
 pub async fn update_state(
